Add fields for social sign up
