--- kobo/apps/accounts/forms/__init__.py
@@ -1,2 +1,0 @@
-from .login import LoginForm
-from .signup import SignupForm
